Add files via upload
Module 5 Project:

A program that modifies the Module 4 project to read a list of integers from a scanned input and save them into a stack sorted from smallest to largest, while making use of the collections framework. 

This folder includes a text file that outlines the plan for switching from a linked list to a stack.